pyproject.toml
[project]
name = "news-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "openai>=2.7.1",
    "playwright>=1.55.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
